+++ src/components/body/Body.js
@@ -1,35 +1,8 @@
 import React from 'react';
 import CardList from './cards/CardList';
-import styled from 'styled-components';
-import Vo from '../../img/vo.png';
-import Nvo from '../../img/nvo.png';
+import BodyHeader from './body-header/BodyHeader';
 import './Body.css';
-
-const Label = styled.label`
-    display: flex;
-    margin: 20px 5px 10px;
-    height: 50px;
-`;
-const HiddenCheckbox = styled.input.attrs({ type: 'checkbox' })`
-    display: none;
-`;
-const Checkmark = styled.div`
-    position: relative;
-    width: 40px;
-    height: 40px;
-    border: 2px solid orangered;
-    background: url(${(props) => (props.imgCheck ? Vo : Nvo)});
-    background-size: cover;
-    background-color: rgba(201, 25, 25, 0.39);
-    border: 2px solid black;
-    cursor: pointer;
-    margin: auto 10px;
-`;
-const TextViewObly = styled.h2`
-    color: white;
-    margin: auto 0;
-    font-weight: bolder;
-`;
+//import { CardContext } from '../../context/CardContext';
 
 class Body extends React.Component {
     constructor() {
@@ -48,20 +21,21 @@ class Body extends React.Component {
                 ? 'ТОЛЬКО ПРОСМОТР'
                 : 'ПРОСМОТР И РЕДАКТИРОВАНИЕ',
         });
+        this.props.changeModeView();
     }
 
     render() {
         return (
+            //<CardContext viewOnly={this.state.viewOnly}>
             <div className="divBody">
-                <div>
-                    <Label>
-                        <HiddenCheckbox onChange={this.changeView} />
-                        <Checkmark imgCheck={this.state.viewOnly} />
-                        <TextViewObly>{this.state.textCheck}</TextViewObly>
-                    </Label>
-                </div>
+                <BodyHeader
+                    changeView={this.changeView}
+                    viewOnly={this.state.viewOnly}
+                    textCheck={this.state.textCheck}
+                />
                 <CardList viewOnly={this.state.viewOnly} />
             </div>
+            //</CardContext>
         );
     }
 }
