+++ src/components/header/Header.css
@@ -1,0 +1,11 @@
+.styleHeader {
+    width: 100%;
+    height: 100;
+    background-color: rgb(145, 145, 145);
+    font-size: 60px;
+    font-weight: bolder;
+    font-family: Comic Sans MS, Arial;
+    color: white;
+    text-align: center;
+    text-shadow: 2px 5px 1px black;
+}
